requirements.txt
langchain
unstructured
unstructured[local-inference]
pinecone-client
openai